refact(e2e): Move environment variables into its own file
